+++ src/store/features/acknowledgments/acknowledgments.type.ts
@@ -1,5 +1,0 @@
-import { IPage } from '../../../types/page'
-
-export interface IAcknowledgmentsState {
-  acknowledgments?: IPage
-}