+++ CHANGELOG.md
@@ -16,14 +16,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Footer link to Management/back office
 - Config for CodeQL
 - Config for Dependabot
+- Status page link in footer
+- feat: add new pages for About, Boards, and Publish
+- feat(#94): update editor labels for Scientific and Editorial boards
+- Transform for author - ethical charter into an independent page
 
 ### Fixed
 - Volume page, display other volumes with a link
+- Fixed truncated markdown not rendered in /HomeSections/PresentationSection
 
 ### Changed
 - Sort editorial board members by role : Editors-in-chief first, then editors
-- Do not display empty sections (absracts, keywords, relationships) 
-
+- Do not display empty sections (abstracts, keywords, relationships) 
+- Refactored language switcher for better UI/UX and A11y
 
 <!--